Updated tests for updated analysis pipeline.
Tests updated and devtools::check() passes.

The updated fit_model() function now checks the template directory for processed model training document RDS files in data > model_graphs. If the processed files already exist, the funciton loads them. Otherwise, the funciton processes the model training documents. The updated analyze_questioned_documents() function does something similar with the quesitoned documents. Because processing handwriting takes several minutes, I didn't want to fit_model() and analyze_questioned_documents() to process any documents during tests so I put the pre-processed RDS files  in tests > testthat > fixtures > template > data > model_graphs and questioned_graphs. These files are not small, so this might not be the best way to handle this, but at least for now that's what I'm going with.

Getting cluster assignments also takes several minutes, so I placed model_clusters.rds and questioned_clusters.rds in tests > testthat > fixtures > template > data. Again, these files are not small, so this might not be a good long term solution.

+++ R/ClusterModeling_modelfunctions.R
@@ -72,11 +72,15 @@ fit_model <- function(template_dir,
   )
 
   # get cluster assignments
-  model_clusters <- get_clusterassignment(
-    clustertemplate = template,
-    input_dir = file.path(main_dir, "data", "model_graphs")
-  )
-  saveRDS(model_clusters, file.path(template_dir, "data", "model_clusters.rds"))
+  if ( file.exists(file.path(template_dir, "data", "model_clusters.rds")) ){
+    model_clusters <- readRDS(file.path(template_dir, "data", "model_clusters.rds"))
+  } else {
+    model_clusters <- get_clusterassignment(
+      clustertemplate = template,
+      input_dir = file.path(main_dir, "data", "model_graphs")
+    )
+    saveRDS(model_clusters, file.path(template_dir, "data", "model_clusters.rds"))
+  }
 
   # format model data
   model_data <- format_model_data(
@@ -279,11 +283,17 @@ analyze_questioned_documents <- function(template_dir, questioned_images_dir, mo
 
   # load template
   template <- readRDS(file.path(template_dir, "data", "template.rds"))
-  questioned_clusters <- get_clusterassignment(
-    clustertemplate = template,
-    input_dir = file.path(main_dir, "data", "questioned_graphs")
-  )
-  saveRDS(questioned_clusters, file.path(template_dir, "data", "questioned_clusters.rds"))
+  
+  # get cluster assignments
+  if ( file.exists(file.path(template_dir, "data", "questioned_clusters.rds")) ){
+    questioned_clusters <- readRDS(file.path(template_dir, "data", "questioned_clusters.rds"))
+  } else {
+    questioned_clusters <- get_clusterassignment(
+      clustertemplate = template,
+      input_dir = file.path(template_dir, "data", "questioned_graphs")
+    )
+    saveRDS(questioned_clusters, file.path(template_dir, "data", "questioned_clusters.rds"))
+  }
 
   # format data
   questioned_data <- format_questioned_data(
